Implemented type selection in the form builder.
Changed registration of events to first delete existing ones (in case of recalling BuildForm(); think loops). Still need to do so in a few more places.
Added WpfToolkit library to get the spinner UI elements.
Changed the elements created for Int and Real types to spinners.
Changed the FormsStmnt to use the TabControl instead.
Made forms scrollable.
Changed the way in which the Least Upper Bound is selected.
Generalized some functions and properties related to types into an interface in Algebra.Core.
Added some random tests that I used to experiment with using Action and Func inside anonymous objects.